+++ include/ast/common/Identifier.hpp
@@ -20,7 +20,7 @@ namespace AST
         Identifier(Lexer::Token &identifier,
                    std::vector<IdentifierArgument> &&arguments);
 
-        void print(std::ostream &os, uint8_t tab) const override;
+        void print(std::ostream &os, uint32_t tab) const override;
     };
 
 } // namespace AST
