+++ src/ast/expr/BinaryExpr.cpp
@@ -17,7 +17,7 @@ namespace AST
                                               : &operation.range.end;
     }
 
-    void BinaryExpr::print(std::ostream &os, uint8_t tab) const
+    void BinaryExpr::print(std::ostream &os, uint32_t tab) const
     {
         std::string indentation = utils::tab(tab - 1, 4),
                     inner_indentation = utils::tab(tab, 4);
