+++ src/ast/stmt/function/FunctionRetStmt.cpp
@@ -12,7 +12,7 @@ namespace AST
             end_position = return_value->end_position;
     }
 
-    void FunctionReturnStmt::print(std::ostream &os, uint8_t tab) const
+    void FunctionReturnStmt::print(std::ostream &os, uint32_t tab) const
     {
         std::string indentation = utils::tab(tab - 1, 4),
                     inner_indentation = utils::tab(tab, 4);
