+++ src/ast/stmt/petal/PetalDeclStmt.cpp
@@ -13,7 +13,7 @@ namespace AST
     {
     }
 
-    void PetalDeclarationStmt::print(std::ostream &os, uint8_t tab) const
+    void PetalDeclarationStmt::print(std::ostream &os, uint32_t tab) const
     {
         std::string indentation = utils::tab(tab - 1, 4),
                     inner_indentation = utils::tab(tab, 4);
