+++ src/ast/stmt/petal/PetalDefStmt.cpp
@@ -14,7 +14,7 @@ namespace AST
         is_definition_ = true;
     }
 
-    void PetalDefinitionStmt::print(std::ostream &os, uint8_t tab) const
+    void PetalDefinitionStmt::print(std::ostream &os, uint32_t tab) const
     {
         std::string indentation = utils::tab(tab - 1, 4),
                     inner_indentation = utils::tab(tab, 4);
