+++ src/ast/type/SuffixedType.cpp
@@ -28,7 +28,7 @@ namespace AST
         return str;
     }
 
-    void SuffixedType::print(std::ostream &os, uint8_t tab) const
+    void SuffixedType::print(std::ostream &os, uint32_t tab) const
     {
         std::string indentation = utils::tab(tab - 1, 4),
                     inner_indentation = utils::tab(tab, 4);
